requirements.txt
python==3.6
numpy>=1.14
numba
sklearn
scipy
tqdm
umap-learn
